+++ .vscode/settings.json
@@ -7,7 +7,6 @@
         "attegg",
         "Atteggs",
         "blurple",
-        "Eggspansions",
         "eggsplode",
         "eggsploded",
         "noped",
