+++ eggsplode/ui/__init__.py
@@ -1,10 +1,17 @@
 """
-This module contains the views for the Eggsplode game UI.
+Contains the views for the Eggsplode game UI.
 """
 
 from .base import BaseView, TextView
 from .nope import NopeView
 from .play import PlayView
 from .selections import SelectionView, ChoosePlayerView, DefuseView
-from .start import EndGameView, StartGameView, SettingsModal, HelpView, InfoView
+from .start import (
+    EndGameView,
+    StartGameView,
+    SettingsModal,
+    HelpView,
+    InfoView,
+    EditRecipeModal,
+)
 from .turn import TurnView
