+++ requirements.txt
@@ -1,5 +1,5 @@
 # py-cord>=2.7.0
-git+https://github.com/Pycord-Development/pycord.git@refs/pull/2707/head
+git+https://github.com/Pycord-Development/pycord.git
 python-dotenv
 psutil
 coverage
