+++ resources/cards.json
@@ -12,14 +12,12 @@
     "nope": {
         "title": "Nope",
         "description": "Stop the action of another card. Can be used by clicking on the **Nope** button.",
-        "emoji": "🛑",
-        "count": 5
+        "emoji": "🛑"
     },
     "attegg": {
         "title": "Attegg",
         "description": "End your turn without drawing and force the next player to take 2 turns in a row.",
         "emoji": "⚡",
-        "count": 4,
         "usable": true,
         "explicit": true
     },
@@ -27,28 +25,24 @@
         "title": "Skip",
         "description": "End your turn without drawing.",
         "emoji": "🏃‍➡️",
-        "count": 4,
         "usable": true
     },
     "shuffle": {
         "title": "Shuffle",
         "description": "Shuffle the deck.",
         "emoji": "🌀",
-        "count": 4,
         "usable": true
     },
-    "predict": {
+    "see_future": {
         "title": "See the future",
         "description": "Look at the first 3 cards on the deck.",
         "emoji": "🔮",
-        "count": 5,
         "usable": true
     },
     "food0": {
         "title": "Boiled egg",
         "description": "Use 2 of this card to steal a card from another player!",
         "emoji": "🥚",
-        "count": 4,
         "usable": true,
         "explicit": true,
         "combo": 1
@@ -57,7 +51,6 @@
         "title": "Fried egg",
         "description": "Use 2 of this card to steal a card from another player!",
         "emoji": "🍳",
-        "count": 4,
         "usable": true,
         "explicit": true,
         "combo": 1
@@ -66,7 +59,6 @@
         "title": "Ham & Eggs",
         "description": "Use 2 of this card to steal a card from another player!",
         "emoji": "🥓",
-        "count": 4,
         "usable": true,
         "explicit": true,
         "combo": 1
@@ -75,7 +67,6 @@
         "title": "Pancake",
         "description": "Use 2 of this card to steal a card from another player!",
         "emoji": "🥞",
-        "count": 4,
         "usable": true,
         "explicit": true,
         "combo": 1
@@ -84,7 +75,6 @@
         "title": "Fried chicken",
         "description": "Use 2 of this card to steal a card from another player!",
         "emoji": "🍗",
-        "count": 4,
         "usable": true,
         "explicit": true,
         "combo": 1
@@ -92,95 +82,73 @@
     "radioeggtive": {
         "title": "Radioeggtive",
         "description": "Put this card back into the deck face up.",
-        "emoji": "☢️",
-        "expansion": "radioeggtive"
+        "emoji": "☢️"
     },
     "radioeggtive_face_up": {
         "title": "Radioeggtive (Face up)!",
         "description": "You lose the game. No escape.",
-        "emoji": "💥",
-        "expansion": "radioeggtive"
+        "emoji": "💥"
     },
     "draw_from_bottom": {
         "title": "Draw from the bottom",
         "description": "End your turn by drawing from the bottom instead of the top.",
         "emoji": "🔽",
-        "count": 4,
-        "usable": true,
-        "expansion": "radioeggtive"
+        "usable": true
     },
     "targeted_attegg": {
         "title": "Targeted Attegg",
         "description": "End your turn without drawing and force a player of your choice to take 2 turns in a row.",
         "emoji": "🎯",
-        "count": 4,
         "usable": true,
-        "explicit": true,
-        "expansion": "radioeggtive"
+        "explicit": true
     },
     "alter_future": {
         "title": "Alter the future",
         "description": "Look at the next 3 cards on the deck and rearrange them in any order.",
         "emoji": "🪄",
-        "count": 4,
-        "usable": true,
-        "expansion": "radioeggtive"
+        "usable": true
     },
     "reverse": {
         "title": "Reverse",
         "description": "End your turn without drawing and reverse the order of play.",
         "emoji": "🔄",
-        "count": 4,
-        "usable": true,
-        "expansion": "radioeggtive"
+        "usable": true
     },
-    "wisecracker": {
+    "eggsperiment": {
         "title": "Wise-Cracker",
         "description": "Make the player with the other Wise-Cracker eggsplode. Play in pair to decide who must eggsplode.",
-        "emoji": "🧠",
-        "count": 0,
+        "emoji": "🧪",
         "usable": true,
-        "explicit": true,
-        "expansion": "wisecracker"
+        "explicit": true
     },
     "bury": {
         "title": "Bury",
         "description": "End your turn by looking at the top card and putting it back into the deck.",
         "emoji": "🪦",
-        "count": 2,
-        "usable": true,
-        "expansion": "wisecracker"
+        "usable": true
     },
     "self_attegg": {
         "title": "Self-Attegg",
         "description": "Play 2 more turns in a row.",
         "emoji": "↩️",
-        "count": 4,
-        "usable": true,
-        "expansion": "wisecracker"
+        "usable": true
     },
     "share_future": {
         "title": "Share the future",
         "description": "Rearrange the next 3 cards on the deck and show them to the next player.",
         "emoji": "👀",
-        "count": 2,
-        "usable": true,
-        "expansion": "wisecracker"
+        "usable": true
     },
     "super_skip": {
         "title": "Super Skip",
         "description": "End your turn without drawing. If you're under the effect of an Attegg, end all of your turns.",
         "emoji": "🚀",
-        "count": 1,
-        "usable": true,
-        "expansion": "wisecracker"
+        "usable": true
     },
     "dig_deeper": {
         "title": "Dig Deeper",
         "description": "End your turn by looking at the top card on the deck. Decide whether to keep it or put it back and blindly draw the next one.",
         "emoji": "⛏️",
-        "count": 4,
-        "usable": true,
-        "expansion": "wisecracker"
+        "usable": true
     }
 }
