--- resources/expansions.json
@@ -1,10 +1,0 @@
-{
-    "radioeggtive": {
-        "name": "Radioeggtive",
-        "emoji": "☢️"
-    },
-    "wisecracker": {
-        "name": "Wise-Cracker (Beta)",
-        "emoji": "🧠"
-    }
-}
