+++ resources/info.json
@@ -1,3 +1,3 @@
 {
-    "version": "1.4"
+    "version": "1.5-development"
 }
