+++ resources/messages.json
@@ -1,7 +1,7 @@
 {
     "action_noped": "🛑 This action has been noped!",
     "advanced_settings": "### Balancing settings",
-    "advanced_settings_description": "🆕 Customize the amount of cards in the deck and turn timers.",
+    "advanced_settings_description": "Customize the amount of cards in the deck and turn timers.",
     "already_in_game": "❌ You are already in the game!",
     "alter_future_placeholder": "{}. card: {}",
     "altered_future": "🪄 <@{}> rearranged the next 3 cards on the deck!",
@@ -26,9 +26,6 @@
     "end_game_button": "Yes, end now",
     "end_game_permission_denied": "❌ You don't have permission to interact with this view!",
     "end_game_warning": "⚠️ **This will end the game and remove all players from it.**\nAre you sure you want to end the game?",
-    "expansions": "### Eggspansions",
-    "expansions_description": "🆕 Eggspansions are additional cards that can be added to the game.",
-    "expansions_updated": "✅ Eggspansions updated!",
     "game_already_exists": "❌ This channel already has a game running!",
     "game_cancelled": "**🚫 Everyone left, so the game has been cancelled. You can start a new game with </start:1325457141628141661>.**",
     "game_ended": "🚫 The current game has been ended manually by <@{}>.",
@@ -141,7 +138,7 @@
     "next_cards": "### Next 3 cards on the deck:",
     "next_turn": "### ⌛ <@{}>'s turn!",
     "no_cards_to_steal": "🎭 {} wanted to steal a card from {}, but they had no cards left. Better luck next time!",
-    "no_expansions": "No eggspansions selected.",
+    "no_recipes": "No recipes selected.",
     "no_nope_cards": "❌ You have no **Nope** cards to play!",
     "no_players_have_cards": "❌ No players have cards!",
     "no_self_nope": "❌ You can't nope your own action!",
@@ -164,6 +161,10 @@
     "predicted": "🔮 <@{}> looked at the next 3 cards on the deck!",
     "radioeggtive": "## ☢️ <@{}> drew the Radioeggtive card and put it back into the deck face up!",
     "radioeggtive_face_up": "## 💥 <@{}> drew the Radioeggtive card face up and eggsploded!",
+    "recipe": "### Recipes",
+    "recipe_description": "🆕 Recipes allow you to customize the cards used in the game. [Learn how to make recipes](https://iqnite.github.io/eggsplode/recipemaker/)",
+    "recipe_json_error": "❌ An error occurred while parsing the recipe's JSON:\n{}\nMake sure the syntax is correct.",
+    "recipe_update_sucess": "✅ Recipe updated successfully",
     "reversed": ":arrows_counterclockwise: <@{}> skipped their turn without drawing and reversed the order of play!",
     "set_status_success": "✅ Status updated!",
     "settings": "## Settings",
@@ -171,8 +172,6 @@
     "settings_updated_error": "❌ {}: {} ({})",
     "settings_updated_success": "✅ {}: {}",
     "shared_future": "👀 <@{}> rearranged the next 3 cards on the deck and allowed <@{}> to view them!",
-    "short_mode": "### Short mode",
-    "short_mode_description": "Removes 50-60% of the cards from the deck. Enabled by default for 2 players.",
     "shuffled": "🌀 <@{}> shuffled the deck!",
     "skipped": "⏩ <@{}> skipped their turn without drawing!",
     "start": "# New game",
@@ -199,8 +198,8 @@
     "user_not_in_game": "❌ You are not in this game!",
     "version_eggsplode": "Eggsplode Version: `{}`",
     "version_pycord": "Pycord Version: `{}`",
-    "wisecracker_defused": "## 🔧 <@{}> used a Wise-Cracker card and forced <@{}> to defuse!",
-    "wisecracker_eggsploded": "## 💥 <@{}> used a Wise-Cracker and made <@{}> eggsplode!",
-    "wisecracker_exposed": "## 🧠 <@{}> tried to use a Wise-Cracker. Unfortunately, nobody else had one, so they took it back into their hand!",
+    "eggsperiment_defused": "## 🔧 <@{}> used a Wise-Cracker card and forced <@{}> to defuse!",
+    "eggsperiment_eggsploded": "## 💥 <@{}> used a Wise-Cracker and made <@{}> eggsplode!",
+    "eggsperiment_exposed": "## 🧪 <@{}> tried to use a Wise-Cracker. Unfortunately, nobody else had one, so they took it back into their hand!",
     "you_drew_card": "You drew: {} {}!"
 }
