+++ resources/recipes.json
@@ -1,0 +1,216 @@
+{
+    "classic": {
+        "name": "Classic",
+        "description": "The classic Eggsplode game",
+        "emoji": "💥",
+        "cards": {
+            "defuse": {
+                "amount": 2,
+                "hand_out": 1,
+                "preserve": true
+            },
+            "nope": 5,
+            "attegg": 4,
+            "skip": 4,
+            "shuffle": 4,
+            "see_future": 5,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4,
+            "food3": 4
+        }
+    },
+    "classic_radioeggtive": {
+        "name": "Radioeggtive Expansion",
+        "description": "The Radioeggtive expansion + the classic game",
+        "emoji": "☢️",
+        "cards": {
+            "defuse": {
+                "amount": 2,
+                "hand_out": 1,
+                "preserve": true
+            },
+            "nope": 5,
+            "attegg": 4,
+            "skip": 4,
+            "shuffle": 4,
+            "see_future": 5,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4,
+            "food3": 4,
+            "radioeggtive": {
+                "amount": 1,
+                "hand_out": 0,
+                "expand_beyond": null,
+                "preserve": true
+            },
+            "draw_from_bottom": 4,
+            "targeted_attegg": 4,
+            "alter_future": 4,
+            "reverse": 4
+        }
+    },
+    "classic_eggsperiment": {
+        "name": "Eggsperiment Expansion",
+        "description": "The Eggsperiment expansion + the classic game",
+        "emoji": "🧪",
+        "cards": {
+            "defuse": {
+                "amount": 2,
+                "hand_out": 1,
+                "preserve": true
+            },
+            "nope": 5,
+            "attegg": 4,
+            "skip": 4,
+            "shuffle": 4,
+            "see_future": 5,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4,
+            "food3": 4,
+            "eggsperiment": {
+                "amount": 2,
+                "hand_out": 0,
+                "expand_beyond": null,
+                "preserve": true
+            },
+            "bury": 2,
+            "self_attegg": 4,
+            "share_future": 2,
+            "super_skip": 1,
+            "dig_deeper": 4
+        }
+    },
+    "mind_games": {
+        "name": "Mind Games",
+        "description": "You know so much, but you can do so little.",
+        "emoji": "🧠",
+        "cards_per_player": 6,
+        "cards": {
+            "defuse": {
+                "hand_out": 1
+            },
+            "bury": 4,
+            "see_future": 6,
+            "skip": 4,
+            "draw_from_bottom": 2,
+            "reverse": 4,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4,
+            "food3": 4
+        }
+    },
+    "card_hoarders": {
+        "name": "Card Hoarders",
+        "description": "No skipping, only drawing.",
+        "emoji": "🗃️",
+        "cards_per_player": 6,
+        "cards": {
+            "defuse": {
+                "hand_out": 1
+            },
+            "radioeggtive": {
+                "amount": 1,
+                "hand_out": 0,
+                "expand_beyond": null,
+                "preserve": true
+            },
+            "see_future": 4,
+            "share_future": 4,
+            "draw_from_bottom": 5,
+            "shuffle": 3,
+            "nope": 4,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4
+        }
+    },
+    "egg_fight": {
+        "name": "Egg fight",
+        "description": "Highly aggressive. Ideal for quick games.",
+        "emoji": "👊",
+        "cards_per_player": 3,
+        "cards": {
+            "defuse": {
+                "hand_out": 1
+            },
+            "attegg": 4,
+            "targeted_attegg": 2,
+            "nope": 3,
+            "alter_future": 2,
+            "reverse": 3
+        }
+    },
+    "attegg_of_atteggs": {
+        "name": "Attegg of the Atteggs",
+        "description": "Stegg those Atteggs!",
+        "emoji": "⚡",
+        "cards": {
+            "defuse": {
+                "amount": 1,
+                "hand_out": 1,
+                "preserve": true
+            },
+            "attegg": 6,
+            "targeted_attegg": 4,
+            "skip": 4,
+            "super_skip": 2,
+            "reverse": 3,
+            "see_future": 3,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4,
+            "nope": 4
+        }
+    },
+    "Power Plant": {
+        "name": "Power Plant",
+        "description": "Radioeggtive only.",
+        "emoji": "☢️",
+        "cards_per_player": 7,
+        "cards": {
+            "radioeggtive": {
+                "auto_amount": -1,
+                "hand_out": 0,
+                "expand_beyond": null,
+                "preserve": true
+            },
+            "skip": 5,
+            "nope": 4,
+            "shuffle": 5,
+            "alter_future": 3,
+            "share_future": 4,
+            "draw_from_bottom": 6,
+            "reverse": 4,
+            "food0": 6,
+            "food1": 6
+        }
+    },
+    "danger_mode": {
+        "name": "Danger Mode",
+        "description": "Higly explosive. Ideal for quick games.",
+        "emoji": "🌩️",
+        "cards": {
+            "eggsplode": {
+                "auto_amount": 2
+            },
+            "defuse": {
+                "amount": 1,
+                "hand_out": 1,
+                "preserve": true
+            },
+            "attegg": 4,
+            "targeted_attegg": 2,
+            "super_skip": 1,
+            "shuffle": 3,
+            "reverse": 5,
+            "draw_from_bottom": 2,
+            "food0": 4,
+            "food1": 4,
+            "food2": 4
+        }
+    }
+}
